refactor: splitter 드래그 로직을 useSplitter 훅으로 분리
ForensicApp 내부의 splitter 마우스 드래그 useEffect와 isDraggingSplitterRef를
useSplitter 훅으로 추출. min/max 경계도 옵션으로 노출.
